move ChatWindowHeader component into ChatWindow component directory
