+++ src/components/ChatWindow/ChatWindowHeader/ChatWindowHeader.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { makeStyles, createStyles } from '@material-ui/core/styles';
+import CloseIcon from '../../../icons/CloseIcon';
+
+import useChatContext from '../../../hooks/useChatContext/useChatContext';
+
+const useStyles = makeStyles(() =>
+  createStyles({
+    container: {
+      height: '56px',
+      background: '#F4F4F6',
+      boxShadow: 'inset 0 -0.1em 0 #E4E7E9',
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      padding: '0 1em',
+    },
+    text: {
+      fontWeight: 'bold',
+    },
+    closeChatWindow: {
+      cursor: 'pointer',
+      display: 'flex',
+    },
+  })
+);
+
+export default function ChatWindowHeader() {
+  const classes = useStyles();
+  const { setIsChatWindowOpen } = useChatContext();
+
+  return (
+    <div className={classes.container}>
+      <div className={classes.text}>Chat</div>
+      <div className={classes.closeChatWindow} onClick={() => setIsChatWindowOpen(false)}>
+        <CloseIcon />
+      </div>
+    </div>
+  );
+}
